chore: :art: rename and move files within client module
